Deleted everything related to platformers and renamed all the needed files
