+++ Assets/Prefabs/Player.prefab
@@ -11,15 +11,18 @@ GameObject:
   - component: {fileID: 1289678593409138194}
   - component: {fileID: 5909287408352961984}
   - component: {fileID: 1289678593409138195}
+  - component: {fileID: 4676619659353751669}
   - component: {fileID: 1289678593409138192}
   - component: {fileID: 1289678593409138193}
   - component: {fileID: 1289678593409138198}
   - component: {fileID: 745240326}
-  - component: {fileID: 4676619659353751669}
-  - component: {fileID: 890480825}
   - component: {fileID: 890480833}
+  - component: {fileID: 745240340}
+  - component: {fileID: 745240351}
+  - component: {fileID: 890480825}
+  - component: {fileID: 1985835157124944172}
   m_Layer: 0
-  m_Name: PlayerTopDown
+  m_Name: Player
   m_TagString: Player
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -34,7 +37,7 @@ Transform:
   m_GameObject: {fileID: 1289678593409138191}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 10, y: 10, z: 1}
+  m_LocalScale: {x: 25, y: 25, z: 0}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1289678594974190353}
@@ -114,6 +117,22 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 4
+--- !u!70 &4676619659353751669
+CapsuleCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1289678593409138191}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0.00007075071, y: -0.0017200708}
+  m_Size: {x: 0.09651344, y: 0.12418039}
+  m_Direction: 0
 --- !u!114 &1289678593409138192
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -172,9 +191,10 @@ MonoBehaviour:
   sceneCamera: {fileID: 0}
   bulletPrefab: {fileID: 5373822962961376425, guid: a8b2aa9be753dfa459d7a5651dfb29f6, type: 3}
   bulletSpawnPoint: {fileID: 1289678594974190353}
-  bulletSpeed: 8
-  fireRate: 0.3
-  bulletLifespan: 5
+  bulletSpeed: 10
+  fireRate: 0.5
+  bulletLifespan: 3
+  bulletStrength: 1
 --- !u!114 &745240326
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -188,24 +208,53 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   healthPoints: 3
+  enemyHealth: 1
   isPlayer: 1
+  hearts:
+  - {fileID: 0}
+  - {fileID: 0}
+  - {fileID: 0}
+  extraHearth: {fileID: 0}
+  extraHearth2: {fileID: 0}
   dieEffect: {fileID: 27824649237413497, guid: 19579fea1d4706d45b403e9fee47c128, type: 3}
---- !u!70 &4676619659353751669
-CapsuleCollider2D:
+--- !u!114 &890480833
+MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1289678593409138191}
-  m_Enabled: 0
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0, y: -0.0049207807}
-  m_Size: {x: 0.9970035, y: 1.994007}
-  m_Direction: 0
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ac72b52eaf5f2d84987af342f76816c0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &745240340
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1289678593409138191}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c1ce8858252bcba47a0a531263e7296f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  neededExperience: 5
+  experienceSlider: {fileID: 0}
+--- !u!114 &745240351
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1289678593409138191}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7c7d6772f2adff849bfcdacaf9643782, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!95 &890480825
 Animator:
   serializedVersion: 5
@@ -227,7 +276,7 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
---- !u!114 &890480833
+--- !u!114 &1985835157124944172
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -236,9 +285,50 @@ MonoBehaviour:
   m_GameObject: {fileID: 1289678593409138191}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ac72b52eaf5f2d84987af342f76816c0, type: 3}
+  m_Script: {fileID: 11500000, guid: 180ecf9b41d478f468eb3e9083753217, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  m_ImpulseDefinition:
+    m_ImpulseChannel: 1
+    m_ImpulseShape: 3
+    m_CustomImpulseShape:
+      serializedVersion: 2
+      m_Curve: []
+      m_PreInfinity: 2
+      m_PostInfinity: 2
+      m_RotationOrder: 4
+    m_ImpulseDuration: 0.2
+    m_ImpulseType: 0
+    m_DissipationRate: 0.25
+    m_RawSignal: {fileID: 0}
+    m_AmplitudeGain: 1
+    m_FrequencyGain: 1
+    m_RepeatMode: 0
+    m_Randomize: 1
+    m_TimeEnvelope:
+      m_AttackShape:
+        serializedVersion: 2
+        m_Curve: []
+        m_PreInfinity: 2
+        m_PostInfinity: 2
+        m_RotationOrder: 4
+      m_DecayShape:
+        serializedVersion: 2
+        m_Curve: []
+        m_PreInfinity: 2
+        m_PostInfinity: 2
+        m_RotationOrder: 4
+      m_AttackTime: 0
+      m_SustainTime: 0.2
+      m_DecayTime: 0.7
+      m_ScaleWithImpact: 1
+      m_HoldForever: 0
+    m_ImpactRadius: 100
+    m_DirectionMode: 0
+    m_DissipationMode: 2
+    m_DissipationDistance: 100
+    m_PropagationSpeed: 343
+  m_DefaultVelocity: {x: 0, y: -1, z: 0}
 --- !u!1 &1289678594974190352
 GameObject:
   m_ObjectHideFlags: 0
