+++ Assets/Scenes/Level 2.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: c387b2945ae058940ba16fc9d5dddb0e
+guid: 5173cc4cd0dfcd044bdfa86c553bd5fc
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
