+++ Assets/Scenes/Level 2.unity.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 02db83ef6ed836547a19ae6a215d1445
-NativeFormatImporter:
+guid: 000469a9d466e58498b115072b30eee2
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
