+++ Assets/Scenes/Level 2/NavMesh-NavMesh.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 67fb49d55a8ed6f4981ff470fd973228
-folderAsset: yes
-DefaultImporter:
+guid: e1d8d7f192da74740b01241ebd62625c
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
