+++ Assets/Scenes/Level 3.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 0456dc733b9a97a4a80730678acb3864
-PrefabImporter:
+guid: 0210de170e5f2544a8df3814f26a0cc8
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
