+++ Assets/Scenes/Level 3/NavMesh-NavMesh.asset.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 2cda990e2423bbf4892e6590ba056729
-DefaultImporter:
+guid: 98e03edda62a5f644b2fe7161972aca9
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
