+++ Assets/Scripts/PlayerMovement.cs
@@ -5,31 +5,22 @@ using UnityEngine.InputSystem;
 
 public class PlayerMovement : MonoBehaviour
 {
-
-    [SerializeField] float runSpeed = 10f;
-    [SerializeField] float jumpSpeed = 5f;
+    [SerializeField] float moveSpeed = 10f;
 
     Vector2 moveInput;
     Rigidbody2D myRigidBody;
     CapsuleCollider2D myBodyCollider;
-    BoxCollider2D myFeetCollider;
     PlayerAnimations _playerAnimations;
-
-
-    float gravityScaleAtStart;
 
 
     void Start()
     {
         myRigidBody = GetComponent<Rigidbody2D>();
         myBodyCollider = GetComponent<CapsuleCollider2D>();
-        myFeetCollider = GetComponent<BoxCollider2D>();
         _playerAnimations = GetComponent<PlayerAnimations>();
-        gravityScaleAtStart = myRigidBody.gravityScale;
-
     }
 
-    void FixedUpdate() 
+    void Update()
     {
         Move();
     }
@@ -38,21 +29,18 @@ public class PlayerMovement : MonoBehaviour
         moveInput = value.Get<Vector2>();
     }
 
-
-    void OnJump(InputValue value){
-        if(!myFeetCollider.IsTouchingLayers(LayerMask.GetMask("Ground"))){
-            return;
-        }
-
-        if(value.isPressed){
-            myRigidBody.velocity += new Vector2 (0f, jumpSpeed);
-        }
+    public void updateSpeed() {
+        moveSpeed += 2;
     }
 
     void Move(){
-        Vector2 playerVelocity = new Vector2(moveInput.x * runSpeed, myRigidBody.velocity.y);
-        _playerAnimations.ChangeState(playerVelocity);
+        Vector2 playerVelocity = moveInput * moveSpeed;
         myRigidBody.velocity = playerVelocity;
+        if(playerVelocity.y != 0 && playerVelocity.x != 0){
+            _playerAnimations.ChangeState(new Vector2(0, playerVelocity.y));
+        }   else {
+            _playerAnimations.ChangeState(playerVelocity);
+        }
     }
 
 }
