--- Assets/Scripts/PlayerMovementTopDown.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eb9c22fe189f0be4a8b64577e245961d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
