+++ Assets/Scripts/PowerUps/IndepencePowerUp.cs
@@ -33,6 +33,6 @@ public class IndepencePowerUp : MonoBehaviour
 
     public void ActivatePowerup()
     {
-        player.GetComponent<PlayerMovementTopDown>().updateSpeed();
+        player.GetComponent<PlayerMovement>().updateSpeed();
     }
 }
