+++ ProjectSettings/EditorBuildSettings.asset
@@ -15,6 +15,6 @@ EditorBuildSettings:
     path: Assets/Scenes/Topdown level 1.unity
     guid: 2cda990e2423bbf4892e6590ba056729
   - enabled: 1
-    path: Assets/Scenes/Topdown level 2.unity
+    path: Assets/Scenes/Level 1.unity
     guid: ce93e5165f35b4b419d6bad5ca6db586
   m_configObjects: {}
